added check for recheck envName

+++ internal/command/commands/env.go
@@ -191,6 +191,14 @@ func (e *Env) Run(args []string) int {
 		}
 		emptyParameters := emptyParameters(map[string]string{"--name": *name})
 		if len(emptyParameters) == 0 {
+			val, err := e.Input.Ask("Please re enter the env name:")
+			if err != nil {
+				e.Logger.Error(err.Error())
+			}
+			if val != *name {
+				e.Logger.Error("Env Name does not match !!")
+				return 1
+			}
 			e.Logger.Info("Environment(" + *name + ") deletion initiated")
 			response, err := envClient.DeleteEnv(*name)
 			if err != nil {
